Use TypeScript for compilation
