+++ kiosk-mode.js
@@ -1,288 +1,1 @@
-class KioskMode {
-  constructor() {
-    window.kioskModeEntities = {};
-    if (this.queryString("clear_km_cache")) this.setCache(["kmHeader", "kmSidebar", "kmOverflow", "kmMenuButton", "kmAccount", "kmSearch"], "false");
-    this.ha = document.querySelector("home-assistant");
-    this.main = this.ha.shadowRoot.querySelector("home-assistant-main").shadowRoot;
-    this.user = this.ha.hass.user;
-    this.llAttempts = 0;
-    this.run();
-    this.entityWatch();
-    new MutationObserver(this.watchDashboards).observe(this.main.querySelector("partial-panel-resolver"), {
-      childList: true,
-    });
-  }
-
-  run(lovelace = this.main.querySelector("ha-panel-lovelace")) {
-    if (this.queryString("disable_km") || !lovelace) return;
-    this.getConfig(lovelace);
-  }
-
-  getConfig(lovelace) {
-    this.llAttempts++;
-    try {
-      const llConfig = lovelace.lovelace.config;
-      const config = llConfig.kiosk_mode || {};
-      this.processConfig(lovelace, config);
-    } catch (e) {
-      if (this.llAttempts < 200) {
-        setTimeout(() => this.getConfig(lovelace), 50);
-      } else {
-        console.log("Lovelace config not found, continuing with default configuration.");
-        console.log(e);
-        this.processConfig(lovelace, {});
-      }
-    }
-  }
-
-  processConfig(lovelace, config) {
-    const dash = this.ha.hass.panelUrl;
-    if (!window.kioskModeEntities[dash]) window.kioskModeEntities[dash] = [];
-    this.hideHeader = this.hideSidebar = this.hideOverflow = this.hideAccount = this.hideSearch = this.hideAssistant = this.ignoreEntity = this.ignoreMobile = false;
-
-    // Retrieve localStorage values & query string options.
-    const queryStringsSet = (
-      this.cached([
-          "kmHeader",
-          "kmSidebar",
-          "kmOverflow",
-          "kmMenuButton",
-          "kmAccount",
-          "kmSearch",
-          "kmAssistant"
-      ]) ||
-      this.queryString([
-          "kiosk",
-          "hide_sidebar",
-          "hide_header",
-          "hide_overflow",
-          "hide_menubutton",
-          "hide_account",
-          "hide_search",
-          "hide_assistant"
-      ])
-    );
-    if (queryStringsSet) {
-      this.hideHeader = this.cached("kmHeader") || this.queryString(["kiosk", "hide_header"]);
-      this.hideSidebar = this.cached("kmSidebar") || this.queryString(["kiosk", "hide_sidebar"]);
-      this.hideOverflow = this.cached("kmOverflow") || this.queryString(["kiosk", "hide_overflow"]);
-      this.hideMenuButton = this.cached("kmMenuButton") || this.queryString(["kiosk", "hide_menubutton"]);
-      this.hideAccount = this.cached("kmAccount") || this.queryString(["kiosk", "hide_account"]);
-      this.hideSearch = this.cached("kmSearch") || this.queryString(["kiosk", "hide_search"]);
-      this.hideAssistant = this.cached("kmAssistant") || this.queryString(["kiosk", "hide_assistant"]);
-    }
-
-    // Use config values only if config strings and cache aren't used.
-    this.hideHeader = queryStringsSet ? this.hideHeader : config.kiosk || config.hide_header;
-    this.hideSidebar = queryStringsSet ? this.hideSidebar : config.kiosk || config.hide_sidebar;
-    this.hideOverflow = queryStringsSet ? this.hideOverflow : config.kiosk || config.hide_overflow;
-    this.hideMenuButton = queryStringsSet ? this.hideMenuButton : config.kiosk || config.hide_menubutton;
-    this.hideAccount = queryStringsSet ? this.hideAccount : config.kiosk || config.hide_account;
-    this.hideSearch = queryStringsSet ? this.hideSearch : config.kiosk || config.hide_search;
-    this.hideAssistant = queryStringsSet ? this.hideAssistant : config.kiosk || config.hide_assistant;
-
-    const adminConfig = this.user.is_admin ? config.admin_settings : config.non_admin_settings;
-    if (adminConfig) this.setOptions(adminConfig);
-
-    if (config.user_settings) {
-      for (let conf of this.array(config.user_settings)) {
-        if (this.array(conf.users).some((x) => x.toLowerCase() == this.user.name.toLowerCase())) this.setOptions(conf);
-      }
-    }
-
-    const mobileConfig = this.ignoreMobile ? null : config.mobile_settings;
-    if (mobileConfig) {
-      const mobileWidth = mobileConfig.custom_width ? mobileConfig.custom_width : 812;
-      if (window.innerWidth <= mobileWidth) this.setOptions(mobileConfig);
-    }
-
-    const entityConfig = this.ignoreEntity ? null : config.entity_settings;
-    if (entityConfig) {
-      for (let conf of entityConfig) {
-        const entity = Object.keys(conf.entity)[0];
-        if (!window.kioskModeEntities[dash].includes(entity)) window.kioskModeEntities[dash].push(entity);
-        if (this.ha.hass.states[entity].state == conf.entity[entity]) {
-          if ("hide_header" in conf) this.hideHeader = conf.hide_header;
-          if ("hide_sidebar" in conf) this.hideSidebar = conf.hide_sidebar;
-          if ("hide_overflow" in conf) this.hideOverflow = conf.hide_overflow;
-          if ("hide_account" in conf) this.hideAccount = conf.hide_account;
-          if ("hide_search" in conf) this.hideSearch = conf.hide_search;
-          if ("hide_assistant" in conf) this.hideAssistant = config.hide_assistant;
-          if ("hide_menubutton" in conf) this.hideMenuButton = conf.hide_menubutton;
-          if ("kiosk" in conf) this.hideHeader = this.hideSidebar = conf.kiosk;
-        }
-      }
-    }
-
-    this.insertStyles(lovelace);
-  }
-
-  insertStyles(lovelace) {
-    const huiRoot = lovelace.shadowRoot.querySelector("hui-root").shadowRoot;
-    const drawerLayout = this.main.querySelector("app-drawer-layout");
-    const appToolbar = huiRoot.querySelector("app-toolbar");
-    const overflowStyle = "app-toolbar > ha-button-menu{display:none !important;}";
-    const sideBarRoot = drawerLayout.querySelector("app-drawer").querySelector("ha-sidebar").shadowRoot;
-    const menuButtonStyle = "ha-menu-button{display:none !important;}";
-    const searchStyle = "app-toolbar > ha-icon-button:first-of-type{display:none !important;}";
-    const assistantStyle = "app-toolbar > ha-icon-button:nth-of-type(2){display:none !important;}";
-    const headerStyle = "#view{min-height:100vh !important;--header-height:0;}app-header{display:none;}";
-  
-    if (this.hideHeader || this.hideOverflow) {
-      const styles = [
-          this.hideHeader ? headerStyle : "",
-          this.hideOverflow ? overflowStyle : ""
-      ];
-      this.addStyle(styles.join(""), huiRoot);
-      if (this.queryString("cache")) {
-        if (this.hideHeader) this.setCache("kmHeader", "true");
-        if (this.hideOverflow) this.setCache("kmOverflow", "true");
-      }
-    } else {
-      this.removeStyle(huiRoot);
-    }
-
-    if (this.hideSidebar) {
-      this.addStyle(":host{--app-drawer-width:0 !important;}#drawer{display:none;}", drawerLayout);
-      if (this.queryString("cache")) this.setCache("kmSidebar", "true");
-    } else {
-      this.removeStyle(drawerLayout);
-    }
-
-    if (
-      this.hideAccount ||
-      this.hideMenuButton
-    ) {
-      const styles = [
-          this.hideAccount ? ".profile{display:none !important;}" : "",
-          this.hideMenuButton ? ".menu ha-icon-button{display:none !important;}" : ""
-      ];
-      this.addStyle(styles.join(""), sideBarRoot);
-      if (this.hideAccount && this.queryString("cache")) this.setCache("kmAccount", "true");
-    } else {
-      this.removeStyle(sideBarRoot);
-    }
-
-    if (
-      this.hideSearch ||
-      this.hideAssistant ||
-      this.hideMenuButton ||
-      this.hideSidebar
-    ) {
-      const styles = [
-          this.hideSearch ? searchStyle : "",
-          this.hideAssistant ? assistantStyle : "",
-          this.hideMenuButton ||  this.hideSidebar ? menuButtonStyle : ""
-      ];
-      this.addStyle(styles.join(""), appToolbar);
-      if (this.queryString("cache")) {
-          if (this.hideSearch) this.setCache("kmSearch", "true");
-          if (this.hideAssistant) this.setCache("kmAssistant", "true");
-          if (this.hideMenuButton) this.setCache("kmMenuButton", "true");
-      }
-    } else {
-      this.removeStyle(appToolbar);
-    }
-
-    // Resize window to "refresh" view.
-    window.dispatchEvent(new Event("resize"));
-
-    this.llAttempts = 0;
-  }
-
-  // Run on dashboard change.
-  watchDashboards(mutations) {
-    mutations.forEach(({ addedNodes }) => {
-      for (let node of addedNodes) if (node.localName == "ha-panel-lovelace") window.KioskMode.run(node);
-    });
-  }
-
-  // Run on entity change.
-  async entityWatch() {
-    (await window.hassConnection).conn.subscribeMessage((e) => this.entityWatchCallback(e), {
-      type: "subscribe_events",
-      event_type: "state_changed",
-    });
-  }
-
-  entityWatchCallback(event) {
-    const entities = window.kioskModeEntities[this.ha.hass.panelUrl] || [];
-    if (
-      entities.length &&
-      event.event_type == "state_changed" &&
-      entities.includes(event.data.entity_id) &&
-      (!event.data.old_state || event.data.new_state.state != event.data.old_state.state)
-    ) {
-      this.run();
-    }
-  }
-
-  setOptions(config) {
-    this.hideHeader = config.kiosk || config.hide_header;
-    this.hideSidebar = config.kiosk || config.hide_sidebar;
-    this.hideOverflow = config.kiosk || config.hide_overflow;
-    this.hideMenuButton = config.kiosk || config.hide_menubutton;
-    this.hideAccount = config.kiosk || config.hide_account;
-    this.hideSearch = config.kiosk || config.hide_search;
-    this.hideAssistant = config.kiosk || config.hide_assistant;
-    this.ignoreEntity = config.ignore_entity_settings;
-    this.ignoreMobile = config.ignore_mobile_settings;
-  }
-
-  // Convert to array.
-  array(x) {
-    return Array.isArray(x) ? x : [x];
-  }
-
-  // Return true if keyword is found in query strings.
-  queryString(keywords) {
-    return this.array(keywords).some((x) => window.location.search.includes(x));
-  }
-
-  // Set localStorage item.
-  setCache(k, v) {
-    this.array(k).forEach((x) => window.localStorage.setItem(x, v));
-  }
-
-  // Retrieve localStorage item as bool.
-  cached(key) {
-    return this.array(key).some((x) => window.localStorage.getItem(x) == "true");
-  }
-
-  styleExists(elem) {
-    return elem.querySelector(`#kiosk_mode_${elem.localName}`);
-  }
-
-  addStyle(css, elem) {
-    if (!this.styleExists(elem)) {
-      const style = document.createElement("style");
-      style.setAttribute("id", `kiosk_mode_${elem.localName}`);
-      style.innerHTML = css;
-      elem.appendChild(style);
-    }
-  }
-
-  removeStyle(elements) {
-    this.array(elements).forEach((elem) => {
-      if (this.styleExists(elem)) elem.querySelector(`#kiosk_mode_${elem.localName}`).remove();
-    });
-  }
-}
-
-// Overly complicated console tag.
-const conInfo = { header: "%c≡ kiosk-mode".padEnd(27), ver: "%cversion *DEV " };
-const br = "%c\n";
-const maxLen = Math.max(...Object.values(conInfo).map((el) => el.length));
-for (const [key] of Object.entries(conInfo)) {
-  if (conInfo[key].length <= maxLen) conInfo[key] = conInfo[key].padEnd(maxLen);
-  if (key == "header") conInfo[key] = `${conInfo[key].slice(0, -1)}⋮ `;
-}
-const header =
-  "display:inline-block;border-width:1px 1px 0 1px;border-style:solid;border-color:#424242;color:white;background:#03a9f4;font-size:12px;padding:4px 4.5px 5px 6px;";
-const info = "border-width:0px 1px 1px 1px;padding:7px;background:white;color:#424242;line-height:0.7;";
-console.info(conInfo.header + br + conInfo.ver, header, "", `${header} ${info}`);
-
-// Initial Run
-Promise.resolve(customElements.whenDefined("hui-view")).then(() => {
-  window.KioskMode = new KioskMode();
-});
+!function(){"use strict";function e(e,t,i,s){return new(i||(i=Promise))((function(n,o){function h(e){try{a(s.next(e))}catch(e){o(e)}}function r(e){try{a(s.throw(e))}catch(e){o(e)}}function a(e){var t;e.done?n(e.value):(t=e.value,t instanceof i?t:new i((function(e){e(t)}))).then(h,r)}a((s=s.apply(e,t||[])).next())}))}function t(e,t){var i,s,n,o,h={label:0,sent:function(){if(1&n[0])throw n[1];return n[1]},trys:[],ops:[]};return o={next:r(0),throw:r(1),return:r(2)},"function"==typeof Symbol&&(o[Symbol.iterator]=function(){return this}),o;function r(r){return function(a){return function(r){if(i)throw new TypeError("Generator is already executing.");for(;o&&(o=0,r[0]&&(h=0)),h;)try{if(i=1,s&&(n=2&r[0]?s.return:r[0]?s.throw||((n=s.return)&&n.call(s),0):s.next)&&!(n=n.call(s,r[1])).done)return n;switch(s=0,n&&(r=[2&r[0],n.value]),r[0]){case 0:case 1:n=r;break;case 4:return h.label++,{value:r[1],done:!1};case 5:h.label++,s=r[1],r=[0];continue;case 7:r=h.ops.pop(),h.trys.pop();continue;default:if(!(n=h.trys,(n=n.length>0&&n[n.length-1])||6!==r[0]&&2!==r[0])){h=0;continue}if(3===r[0]&&(!n||r[1]>n[0]&&r[1]<n[3])){h.label=r[1];break}if(6===r[0]&&h.label<n[1]){h.label=n[1],n=r;break}if(n&&h.label<n[2]){h.label=n[2],h.ops.push(r);break}n[2]&&h.ops.pop(),h.trys.pop();continue}r=t.call(e,h)}catch(e){r=[6,e],s=0}finally{i=n=0}if(5&r[0])throw r[1];return{value:r[0]?r[1]:void 0,done:!0}}([r,a])}}}for(var i=function(){function i(){console.log("TEST TEST"),window.kioskModeEntities={},this.queryString("clear_km_cache")&&this.setCache(["kmHeader","kmSidebar","kmOverflow","kmMenuButton","kmAccount","kmSearch"],"false"),this.ha=document.querySelector("home-assistant"),this.main=this.ha.shadowRoot.querySelector("home-assistant-main").shadowRoot,this.user=this.ha.hass.user,this.llAttempts=0,this.run(),this.entityWatch(),new MutationObserver(this.watchDashboards).observe(this.main.querySelector("partial-panel-resolver"),{childList:!0})}return i.prototype.run=function(e){void 0===e&&(e=this.main.querySelector("ha-panel-lovelace")),!this.queryString("disable_km")&&e&&this.getConfig(e)},i.prototype.getConfig=function(e){var t=this;this.llAttempts++;try{var i=e.lovelace.config.kiosk_mode||{};this.processConfig(e,i)}catch(i){this.llAttempts<200?setTimeout((function(){return t.getConfig(e)}),50):(console.log("Lovelace config not found, continuing with default configuration."),console.log(i),this.processConfig(e,{}))}},i.prototype.processConfig=function(e,t){var i=this,s=this.ha.hass.panelUrl;window.kioskModeEntities[s]||(window.kioskModeEntities[s]=[]),this.hideHeader=this.hideSidebar=this.hideOverflow=this.hideAccount=this.hideSearch=this.hideAssistant=this.ignoreEntity=this.ignoreMobile=!1;var n=this.cached(["kmHeader","kmSidebar","kmOverflow","kmMenuButton","kmAccount","kmSearch","kmAssistant"])||this.queryString(["kiosk","hide_sidebar","hide_header","hide_overflow","hide_menubutton","hide_account","hide_search","hide_assistant"]);n&&(this.hideHeader=this.cached("kmHeader")||this.queryString(["kiosk","hide_header"]),this.hideSidebar=this.cached("kmSidebar")||this.queryString(["kiosk","hide_sidebar"]),this.hideOverflow=this.cached("kmOverflow")||this.queryString(["kiosk","hide_overflow"]),this.hideMenuButton=this.cached("kmMenuButton")||this.queryString(["kiosk","hide_menubutton"]),this.hideAccount=this.cached("kmAccount")||this.queryString(["kiosk","hide_account"]),this.hideSearch=this.cached("kmSearch")||this.queryString(["kiosk","hide_search"]),this.hideAssistant=this.cached("kmAssistant")||this.queryString(["kiosk","hide_assistant"])),this.hideHeader=n?this.hideHeader:t.kiosk||t.hide_header,this.hideSidebar=n?this.hideSidebar:t.kiosk||t.hide_sidebar,this.hideOverflow=n?this.hideOverflow:t.kiosk||t.hide_overflow,this.hideMenuButton=n?this.hideMenuButton:t.kiosk||t.hide_menubutton,this.hideAccount=n?this.hideAccount:t.kiosk||t.hide_account,this.hideSearch=n?this.hideSearch:t.kiosk||t.hide_search,this.hideAssistant=n?this.hideAssistant:t.kiosk||t.hide_assistant;var o=this.user.is_admin?t.admin_settings:t.non_admin_settings;if(o&&this.setOptions(o),t.user_settings)for(var h=0,r=this.array(t.user_settings);h<r.length;h++){var a=r[h];this.array(a.users).some((function(e){return e.toLowerCase()==i.user.name.toLowerCase()}))&&this.setOptions(a)}var d=this.ignoreMobile?null:t.mobile_settings;if(d){var c=d.custom_width?d.custom_width:812;window.innerWidth<=c&&this.setOptions(d)}var u=this.ignoreEntity?null:t.entity_settings;if(u)for(var l=0,p=u;l<p.length;l++){a=p[l];var k=Object.keys(a.entity)[0];window.kioskModeEntities[s].includes(k)||window.kioskModeEntities[s].push(k),this.ha.hass.states[k].state==a.entity[k]&&("hide_header"in a&&(this.hideHeader=a.hide_header),"hide_sidebar"in a&&(this.hideSidebar=a.hide_sidebar),"hide_overflow"in a&&(this.hideOverflow=a.hide_overflow),"hide_account"in a&&(this.hideAccount=a.hide_account),"hide_search"in a&&(this.hideSearch=a.hide_search),"hide_assistant"in a&&(this.hideAssistant=t.hide_assistant),"hide_menubutton"in a&&(this.hideMenuButton=a.hide_menubutton),"kiosk"in a&&(this.hideHeader=this.hideSidebar=a.kiosk))}this.insertStyles(e)},i.prototype.insertStyles=function(e){var t=e.shadowRoot.querySelector("hui-root").shadowRoot,i=this.main.querySelector("app-drawer-layout"),s=t.querySelector("app-toolbar"),n=i.querySelector("app-drawer").querySelector("ha-sidebar").shadowRoot;if(this.hideHeader||this.hideOverflow){var o=[this.hideHeader?"#view{min-height:100vh !important;--header-height:0;}app-header{display:none;}":"",this.hideOverflow?"app-toolbar > ha-button-menu{display:none !important;}":""];this.addStyle(o.join(""),t),this.queryString("cache")&&(this.hideHeader&&this.setCache("kmHeader","true"),this.hideOverflow&&this.setCache("kmOverflow","true"))}else this.removeStyle(t);if(this.hideSidebar?(this.addStyle(":host{--app-drawer-width:0 !important;}#drawer{display:none;}",i),this.queryString("cache")&&this.setCache("kmSidebar","true")):this.removeStyle(i),this.hideAccount||this.hideMenuButton){o=[this.hideAccount?".profile{display:none !important;}":"",this.hideMenuButton?".menu ha-icon-button{display:none !important;}":""];this.addStyle(o.join(""),n),this.hideAccount&&this.queryString("cache")&&this.setCache("kmAccount","true")}else this.removeStyle(n);if(this.hideSearch||this.hideAssistant||this.hideMenuButton||this.hideSidebar){o=[this.hideSearch?"app-toolbar > ha-icon-button:first-of-type{display:none !important;}":"",this.hideAssistant?"app-toolbar > ha-icon-button:nth-of-type(2){display:none !important;}":"",this.hideMenuButton||this.hideSidebar?"ha-menu-button{display:none !important;}":""];this.addStyle(o.join(""),s),this.queryString("cache")&&(this.hideSearch&&this.setCache("kmSearch","true"),this.hideAssistant&&this.setCache("kmAssistant","true"),this.hideMenuButton&&this.setCache("kmMenuButton","true"))}else this.removeStyle(s);window.dispatchEvent(new Event("resize")),this.llAttempts=0},i.prototype.watchDashboards=function(e){e.forEach((function(e){for(var t=0,i=e.addedNodes;t<i.length;t++){var s=i[t];"ha-panel-lovelace"==s.localName&&window.KioskMode.run(s)}}))},i.prototype.entityWatch=function(){return e(this,void 0,void 0,(function(){var e=this;return t(this,(function(t){switch(t.label){case 0:return[4,window.hassConnection];case 1:return t.sent().conn.subscribeMessage((function(t){return e.entityWatchCallback(t)}),{type:"subscribe_events",event_type:"state_changed"}),[2]}}))}))},i.prototype.entityWatchCallback=function(e){var t=window.kioskModeEntities[this.ha.hass.panelUrl]||[];!t.length||"state_changed"!=e.event_type||!t.includes(e.data.entity_id)||e.data.old_state&&e.data.new_state.state==e.data.old_state.state||this.run()},i.prototype.setOptions=function(e){this.hideHeader=e.kiosk||e.hide_header,this.hideSidebar=e.kiosk||e.hide_sidebar,this.hideOverflow=e.kiosk||e.hide_overflow,this.hideMenuButton=e.kiosk||e.hide_menubutton,this.hideAccount=e.kiosk||e.hide_account,this.hideSearch=e.kiosk||e.hide_search,this.hideAssistant=e.kiosk||e.hide_assistant,this.ignoreEntity=e.ignore_entity_settings,this.ignoreMobile=e.ignore_mobile_settings},i.prototype.array=function(e){return Array.isArray(e)?e:[e]},i.prototype.queryString=function(e){return this.array(e).some((function(e){return window.location.search.includes(e)}))},i.prototype.setCache=function(e,t){this.array(e).forEach((function(e){return window.localStorage.setItem(e,t)}))},i.prototype.cached=function(e){return this.array(e).some((function(e){return"true"==window.localStorage.getItem(e)}))},i.prototype.styleExists=function(e){return e.querySelector("#kiosk_mode_".concat(e.localName))},i.prototype.addStyle=function(e,t){if(!this.styleExists(t)){var i=document.createElement("style");i.setAttribute("id","kiosk_mode_".concat(t.localName)),i.innerHTML=e,t.appendChild(i)}},i.prototype.removeStyle=function(e){var t=this;this.array(e).forEach((function(e){t.styleExists(e)&&e.querySelector("#kiosk_mode_".concat(e.localName)).remove()}))},i}(),s={header:"%c≡ kiosk-mode".padEnd(27),ver:"%cversion *DEV "},n=Math.max.apply(Math,Object.values(s).map((function(e){return e.length}))),o=0,h=Object.entries(s);o<h.length;o++){var r=h[o][0];s[r].length<=n&&(s[r]=s[r].padEnd(n)),"header"==r&&(s[r]="".concat(s[r].slice(0,-1),"⋮ "))}var a="display:inline-block;border-width:1px 1px 0 1px;border-style:solid;border-color:#424242;color:white;background:#03a9f4;font-size:12px;padding:4px 4.5px 5px 6px;";console.info(s.header+"%c\n"+s.ver,a,"","".concat(a," ").concat("border-width:0px 1px 1px 1px;padding:7px;background:white;color:#424242;line-height:0.7;")),Promise.resolve(customElements.whenDefined("hui-view")).then((function(){window.KioskMode=new i}))}();
