+++ package.json
@@ -7,17 +7,12 @@
   "author": "Steven \"NemesisRE\" Koeberich <nemesissre@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "babel": "babel kiosk-mode.js --out-file kiosk-mode.js",
-    "browserify": "browserify kiosk-mode.js -o kiosk-mode.js",
-    "uglify": "uglifyjs kiosk-mode.js -b beautify=false,max_line_len=150 --compress --mangle toplevel --output kiosk-mode.js",
-    "build": "yarn babel && yarn browserify && yarn uglify"
+    "build": "rollup --config rollup.config.js --bundleConfigAsCjs"
   },
   "devDependencies": {
-    "@babel/cli": "^7.21.0",
-    "@babel/core": "^7.21.0",
-    "@babel/plugin-transform-runtime": "^7.21.0",
-    "@babel/preset-env": "^7.20.2",
-    "browserify": "^17.0.0",
-    "uglify-js": "^3.17.4"
+    "rollup": "^3.19.1",
+    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-ts": "^3.2.0",
+    "typescript": "^4.9.5"
   }
 }
